extract CalculateDurationFromTimestamp, quick unit test
